requirements.txt
chainlit
langchain-groq
langchain==0.3.0
python-dotenv
-e .